requirements.txt
django-registration
south
django-social-auth
feincms==1.5.3